+++ assets/calamares/install-system.sh
@@ -7,11 +7,15 @@
 sudo mv /etc/fstab /etc/fstab.orig.calamares
 
 # Allow Calamares to scale the window for hidpi displays
-# This is fixed in the Calamares 3.3.0 series
+# This is fixed in the Calamares 3.3.0 series, so we can remove this
+# once we switch to that
+# Upstream commit that will make this obsolete:
+#     https://github.com/calamares/calamares/commit/e9f011b686a0982fb7828e8ac02a8e0784d3b11f
+# Upstream bug:
+#     https://github.com/calamares/calamares/issues/1945
+# Debian bug:
+#     https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=992162
 export QT_AUTO_SCREEN_SCALE_FACTOR=1
-
-# Gnome... not work
-# export QT_QPA_PLATFORM=wayland
 
 # Access control to run calamares as root for xwayland
 xhost +si:localuser:root